implented faster blur algorithm

+++ src/cairo++.hpp
@@ -226,7 +226,7 @@ cairo_pattern_t* cairo_pattern_create_rect_color(cairo_rectangle_t rect,
                                                         double r2, double g2, double b2, double a2,
                                                         double r3, double g3, double b3, double a3);
 
-void cairo_image_surface_blur(cairo_surface_t* surface, double blur_h, double blur_v);
+void cairo_image_surface_blur(cairo_surface_t* surface, float blur_h, float blur_v);
 
 void cairo_apply_matrix(cairo_t* ctx, cairo_matrix_t* mat);
